move reused files to src, create pyproject.toml and fix imports
