+++ pyproject.toml
@@ -1,0 +1,19 @@
+[build-system]
+requires = ["setuptools >= 77.0.3"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "aurora_hpc"
+version = "0.1.0"
+requires-python = ">=3.11"
+dependencies = [
+    "microsoft-aurora",
+    "xarray",
+    "cdsapi",
+    "ipython",
+]
+
+[project.optional-dependencies]
+dev = [
+    "pre-commit",
+]
