+++ train/scripts/timing_data_transfer.py
@@ -6,8 +6,9 @@ import time
 from pathlib import Path
 
 import torch
-from dataset import AuroraDataset, aurora_collate_fn
 from torch.utils.data import DataLoader
+
+from aurora_hpc.dataset import AuroraDataset, aurora_collate_fn
 
 parser = argparse.ArgumentParser()
 parser.add_argument("--xpu", action="store_true", help="boolean of whether to use xpu")
